Move refactored stuff to `protocol/mod.rs`
