+++ src/protocol/mod.rs
@@ -26,9 +26,6 @@ pub(crate) mod send_error;
 pub(crate) mod state;
 pub(crate) mod types;
 
-mod new;
-pub(crate) use self::new::{NewSignaling, NewInitiatorSignaling, NewResponderSignaling};
-
 use self::context::{PeerContext, ServerContext, InitiatorContext, ResponderContext};
 pub(crate) use self::cookie::{Cookie};
 use self::messages::{Message, ServerHello, ServerAuth, ClientHello, ClientAuth, NewResponder};
@@ -36,307 +33,74 @@ use self::messages::{SendError, Token, Key, InitiatorAuthBuilder};
 pub(crate) use self::nonce::{Nonce};
 pub use self::types::{Role};
 pub(crate) use self::types::{HandleAction};
-use self::types::{ClientIdentity, Address};
+use self::types::{Identity, ClientIdentity, Address};
 use self::state::{SignalingState, ServerHandshakeState};
 use self::state::{InitiatorHandshakeState, ResponderHandshakeState};
 
 
-/// The signaling implementation.
-///
-/// This enum contains all the signaling logic. Role specific logic is
-/// delegated to the inner signaling type.
-pub(crate) enum Signaling {
-    Initiator(InitiatorSignaling),
-    Responder(ResponderSignaling),
-}
+/// The main signaling trait.
+pub(crate) trait Signaling {
+    /// Return reference to the common data.
+    fn common(&self) -> &Common;
 
-impl From<InitiatorSignaling> for Signaling {
-    fn from(val: InitiatorSignaling) -> Self {
-        Signaling::Initiator(val)
-    }
-}
+    /// Return mutable reference to the common data.
+    fn common_mut(&mut self) -> &mut Common;
 
-impl From<ResponderSignaling> for Signaling {
-    fn from(val: ResponderSignaling) -> Self {
-        Signaling::Responder(val)
-    }
-}
-
-/// Make it possible to reference the enum variants directly
-use Signaling::{Initiator, Responder};
-
-/// Macro to simplify repetitive enum matching.
-macro_rules! on_inner {
-    ($self:expr, $var:pat, $expr:expr) => {{
-        match *$self {
-            Signaling::Initiator($var) => $expr,
-            Signaling::Responder($var) => $expr,
-        }
-    }}
-}
-
-impl Signaling {
-    /// Create a new initiator signaling instance.
-    pub(crate) fn new_initiator(permanent_key: KeyStore) -> Self {
-        Signaling::Initiator(InitiatorSignaling::new(permanent_key))
-    }
-
-    /// Create a new responder signaling instance.
-    pub(crate) fn new_responder(permanent_key: KeyStore,
-                                initiator_pubkey: PublicKey,
-                                auth_token: Option<AuthToken>) -> Self {
-        Signaling::Responder(ResponderSignaling::new(permanent_key, initiator_pubkey, auth_token))
-    }
-
-    /// Return our role, either initiator or responder.
-    pub(crate) fn role(&self) -> Role {
-        match *self {
-            Signaling::Initiator(_) => Role::Initiator,
-            Signaling::Responder(_) => Role::Responder,
-        }
-    }
-
-    /// Return the signaling state.
-    fn signaling_state(&self) -> SignalingState {
-        on_inner!(self, ref s, s.signaling_state)
-    }
-
-    /// Update the signaling state.
-    fn set_signaling_state(&mut self, state: SignalingState) -> SignalingResult<()> {
-        macro_rules! update { () => {{
-            trace!("Signaling state transition: {:?} -> {:?}", self.signaling_state(), state);
-            on_inner!(self, ref mut s, s.signaling_state = state);
-            Ok(())
-        }} }
-        macro_rules! ignore { () => {{
-            trace!("Ignoring signaling state transition: {:?} -> {:?}", self.signaling_state(), state);
-            Ok(())
-        }} };
-        macro_rules! fail { () => {{
-            Err(SignalingError::InvalidStateTransition(
-                format!("Signaling state: {:?} -> {:?}", self.signaling_state(), state)
-            ))
-        }} };
-
-        match self.signaling_state() {
-            SignalingState::ServerHandshake => match state {
-                SignalingState::ServerHandshake => ignore!(),
-                _ => update!(),
-            },
-            SignalingState::PeerHandshake => {
-                match state {
-                    SignalingState::ServerHandshake => fail!(),
-                    SignalingState::PeerHandshake => ignore!(),
-                    _ => update!(),
-                }
-            },
-            SignalingState::Task => fail!(),
-        }
-    }
-
-    /// Return our assigned client identity.
-    fn identity(&self) -> ClientIdentity {
-        on_inner!(self, ref s, s.identity)
-    }
-
-    /// Set the client identity.
-    fn set_identity(&mut self, identity: ClientIdentity) {
-        on_inner!(self, ref mut s, s.identity = identity);
-    }
-
-    /// Return our permanent keypair.
-    fn permanent_key(&self) -> &KeyStore {
-        on_inner!(self, ref s, &s.permanent_key)
-    }
-
-    /// Return our auth token.
-    pub(crate) fn auth_token(&self) -> Option<&AuthToken> {
-        on_inner!(self, ref s, s.auth_token.as_ref())
-    }
-
-    /// Return the server context.
+    /// Return reference to the server context.
     fn server(&self) -> &ServerContext {
-        on_inner!(self, ref s, &s.server)
+        &self.common().server
     }
 
-    /// Return the mutable server context.
+    /// Return mutable reference to the server context.
     fn server_mut(&mut self) -> &mut ServerContext {
-        on_inner!(self, ref mut s, &mut s.server)
+        &mut self.common_mut().server
     }
 
-    /// Handle an incoming message.
-    pub(crate) fn handle_message(&mut self, bbox: ByteBox) -> SignalingResult<Vec<HandleAction>> {
-        // Validate the nonce
-        match self.validate_nonce(&bbox.nonce) {
-            // It's valid! Carry on.
-            ValidationResult::Ok => {},
-
-            // Drop and ignore some of the messages
-            ValidationResult::DropMsg(warning) => {
-                warn!("Invalid nonce: {}", warning);
-                return Ok(vec![]);
-            },
-
-            // Nonce is invalid, fail the signaling
-            ValidationResult::Fail(reason) => return Err(SignalingError::InvalidNonce(reason)),
-
-            // A critical error occurred
-            ValidationResult::Crash(reason) => return Err(SignalingError::Crash(reason)),
-        }
-
-        // Decode message
-        let obox: OpenBox = if bbox.nonce.source().is_server() {
-            self.decode_server_msg(bbox)?
-        } else {
-            on_inner!(self, ref s, s.decode_peer_msg(bbox))?
-        };
-
-        match self.signaling_state() {
-            SignalingState::ServerHandshake =>
-                self.handle_server_message(obox),
-
-            SignalingState::PeerHandshake if obox.nonce.source().is_server() =>
-                self.handle_server_message(obox),
-            SignalingState::PeerHandshake =>
-                match *self {
-                    Signaling::Initiator(ref mut sig) => sig.handle_peer_message(obox),
-                    Signaling::Responder(ref mut sig) => sig.handle_peer_message(obox),
-                },
-
-            SignalingState::Task =>
-                unimplemented!("TODO: Handle task messages"),
-        }
+    /// Return the identity.
+    fn identity(&self) -> ClientIdentity {
+        self.common().identity
     }
 
-    /// Determine the next server handshake state based on the incoming
-    /// server-to-client message and the current state.
+    /// Return the role.
+    fn role(&self) -> Role {
+        self.common().role
+    }
+
+    /// Return the auth token.
+    fn auth_token(&self) -> Option<&AuthToken> {
+        self.common().auth_token.as_ref()
+    }
+
+    /// Return the server handshake state
+    /// TODO: Remove?
+    fn server_handshake_state(&self) -> ServerHandshakeState {
+        self.server().handshake_state()
+    }
+
+    fn validate_nonce<'a>(&'a mut self, nonce: &Nonce) -> Result<(), ValidationError> {
+        self.validate_nonce_destination(nonce)?;
+        self.validate_nonce_source(nonce)?;
+        self.validate_nonce_csn(nonce)?;
+        self.validate_nonce_cookie(nonce)?;
+        Ok(())
+    }
+
+    /// Return the peer context.
     ///
-    /// This method call may have some side effects, like updates in the peer
-    /// context (cookie, CSN, etc).
-    fn handle_server_message(&mut self, obox: OpenBox) -> SignalingResult<Vec<HandleAction>> {
-        let old_state = self.server().handshake_state().clone();
-        match (old_state, obox.message) {
-            // Valid state transitions
-            (ServerHandshakeState::New, Message::ServerHello(msg)) =>
-                self.handle_server_hello(msg),
-            (ServerHandshakeState::ClientInfoSent, Message::ServerAuth(msg)) =>
-                self.handle_server_auth(msg),
-            (ServerHandshakeState::Done, Message::NewResponder(msg)) =>
-                on_inner!(self, ref mut s, s.handle_new_responder(msg)),
-            (ServerHandshakeState::Done, Message::DropResponder(_msg)) =>
-                unimplemented!("Handling DropResponder messages not yet implemented"),
-            (ServerHandshakeState::Done, Message::SendError(msg)) =>
-                self.handle_send_error(msg),
+    /// May return `None` if the peer is not yet set.
+    fn get_peer(&self) -> Option<&PeerContext>;
 
-            // Any undefined state transition results in an error
-            (s, message) => Err(SignalingError::InvalidStateTransition(
-                format!("Got {} message from server in {:?} state", message.get_type(), s)
-            )),
-        }
-    }
+    /// Return the peer context with the specified address.
+    fn get_peer_with_address_mut(&mut self, addr: Address) -> Option<&mut PeerContext>;
 
-    /// Validate the nonce
-    fn validate_nonce(&mut self, nonce: &Nonce) -> ValidationResult {
-		// A client MUST check that the destination address targets its
-		// assigned identity (or `0x00` during authentication).
-        if self.identity() == ClientIdentity::Unknown
-                && !nonce.destination().is_unknown()
-                && self.server().handshake_state() != ServerHandshakeState::New {
-            // The first message received with a destination address different
-            // to `0x00` SHALL be accepted as the client's assigned identity.
-            // However, the client MUST validate that the identity fits its
-            // role – initiators SHALL ONLY accept `0x01` and responders SHALL
-            // ONLY an identity from the range `0x02..0xff`. The identity MUST
-            // be stored as the client's assigned identity.
-            match self.role() {
-                Role::Initiator => {
-                    if nonce.destination().is_initiator() {
-                        self.set_identity(ClientIdentity::Initiator);
-                        debug!("Assigned identity: {}", &self.identity());
-                    } else {
-                        let msg = format!("cannot assign address {} to a client with role {}", nonce.destination(), self.role());
-                        return ValidationResult::Fail(msg);
-                    }
-                },
-                Role::Responder => {
-                    if nonce.destination().is_responder() {
-                        self.set_identity(ClientIdentity::Responder(nonce.destination().0));
-                        debug!("Assigned identity: {}", &self.identity());
-                    } else {
-                        let msg = format!("cannot assign address {} to a client with role {}", nonce.destination(), self.role());
-                        return ValidationResult::Fail(msg);
-                    }
-                },
-            };
-        }
-        if nonce.destination() != self.identity().into() {
-            let msg = format!("Bad destination: {} (our identity is {})", nonce.destination(), self.identity());
-            return ValidationResult::Fail(msg);
-        }
+    /// Validate the nonce destination
+    fn validate_nonce_destination(&mut self, nonce: &Nonce) -> Result<(), ValidationError>;
 
-        // An initiator SHALL ONLY process messages from the server (0x00). As
-        // soon as the initiator has been assigned an identity, it MAY ALSO accept
-        // messages from other responders (0x02..0xff). Other messages SHALL be
-        // discarded and SHOULD trigger a warning.
-        //
-        // A responder SHALL ONLY process messages from the server (0x00). As soon
-        // as the responder has been assigned an identity, it MAY ALSO accept
-        // messages from the initiator (0x01). Other messages SHALL be discarded
-        // and SHOULD trigger a warning.
-        match nonce.source() {
-            // From server
-            Address(0x00) => {},
+    /// Validate the nonce source
+    fn validate_nonce_source(&mut self, nonce: &Nonce) -> Result<(), ValidationError>;
 
-            // From initiator
-            Address(0x01) => {
-                match self.identity() {
-                    // We're the responder: OK
-                    ClientIdentity::Responder(_) => {},
-                    // Otherwise: Not OK
-                    _ => {
-                        let msg = format!("bad source: {} (our identity is {})", nonce.source(), self.identity());
-                        return ValidationResult::DropMsg(msg);
-                    },
-                }
-            },
-
-            // From responder
-            Address(0x02...0xff) => {
-                match self.identity() {
-                    // We're the initiator: OK
-                    ClientIdentity::Initiator => {},
-                    // Otherwise: Not OK
-                    _ => {
-                        let msg = format!("bad source: {} (our identity is {})", nonce.source(), self.identity());
-                        return ValidationResult::DropMsg(msg);
-                    },
-                }
-            },
-
-            // Required due to https://github.com/rust-lang/rfcs/issues/1550
-            Address(_) => unreachable!(),
-        };
-
-        // Find peer
-        // TODO: Also consider signaling state, see InitiatorSignaling.java getPeerWithId
-        let peer: &mut PeerContext = match nonce.source().0 {
-            0x00 => self.server_mut(),
-            0x01 => match *self {
-                Initiator(_) => return ValidationResult::Crash("Got message from invalid sender that wasn't dropped".into()),
-                Responder(ref mut s) => &mut s.initiator,
-            },
-            addr @ 0x02...0xff => match *self {
-                Initiator(ref mut s) => match s.responders.get_mut(&nonce.source()) {
-                    Some(responder) => responder,
-                    None => return ValidationResult::Fail(format!("could not find responder with address {}", addr)),
-                },
-                Responder(_) => return ValidationResult::Crash("Got message from invalid sender that wasn't dropped".into()),
-            },
-            _ => unreachable!(),
-        };
-
-        let peer_identity = peer.identity();
-
+    /// Validate the nonce CSN
+    fn validate_nonce_csn(&mut self, nonce: &Nonce) -> Result<(), ValidationError> {
         // Validate CSN
         //
         // In case this is the first message received from the sender, the peer:
@@ -352,37 +116,50 @@ impl Signaling {
         //
         // * MUST check that the combined sequence number of the source peer
         //   has been increased by 1 and has not reset to 0.
-        {
-            let mut csn_pair = peer.csn_pair().borrow_mut();
-
-            // If we already have the CSN of the peer,
-            // ensure that it has been increased properly.
-            if let Some(ref mut csn) = csn_pair.theirs {
-                let previous = csn;
-                let current = nonce.csn();
-                if current < previous {
-                    let msg = format!("{} CSN is lower than last time", peer_identity);
-                    return ValidationResult::Fail(msg);
-                } else if current == previous {
-                    let msg = format!("{} CSN hasn't been incremented", peer_identity);
-                    return ValidationResult::Fail(msg);
-                } else {
-                    *previous = current.clone();
-                }
+        let role = self.role();
+        let peer: &mut PeerContext = self.get_peer_with_address_mut(nonce.source()).ok_or_else(|| {
+            if role == Role::Initiator && nonce.source().is_responder() {
+                ValidationError::Fail(format!("Could not find responder with address {}", nonce.source()))
+            } else {
+                ValidationError::Crash("Got message from invalid sender that wasn't dropped".into())
             }
+        })?;
 
-            // Otherwise, this is the first message from that peer.
-            if csn_pair.theirs.is_none() {
-                // Validate the overflow number...
-                if nonce.csn().overflow_number() != 0 {
-                    let msg = format!("First message from {} must have set the overflow number to 0", peer_identity);
-                    return ValidationResult::Fail(msg);
-                }
-                // ...and store the CSN.
-                csn_pair.theirs = Some(nonce.csn().clone());
+        let peer_identity = peer.identity();
+        let mut csn_pair = peer.csn_pair().borrow_mut();
+
+        // If we already have the CSN of the peer,
+        // ensure that it has been increased properly.
+        if let Some(ref mut csn) = csn_pair.theirs {
+            let previous = csn;
+            let current = nonce.csn();
+            if current < previous {
+                let msg = format!("{} CSN is lower than last time", peer_identity);
+                return Err(ValidationError::Fail(msg));
+            } else if current == previous {
+                let msg = format!("{} CSN hasn't been incremented", peer_identity);
+                return Err(ValidationError::Fail(msg));
+            } else {
+                *previous = current.clone();
             }
         }
 
+        // Otherwise, this is the first message from that peer.
+        if csn_pair.theirs.is_none() {
+            // Validate the overflow number...
+            if nonce.csn().overflow_number() != 0 {
+                let msg = format!("First message from {} must have set the overflow number to 0", peer.identity());
+                return Err(ValidationError::Fail(msg));
+            }
+            // ...and store the CSN.
+            csn_pair.theirs = Some(nonce.csn().clone());
+        }
+
+        Ok(())
+    }
+
+    /// Validate the nonce cookie
+    fn validate_nonce_cookie(&mut self, nonce: &Nonce) -> Result<(), ValidationError> {
         // Validate cookie
         //
         // In case this is the first message received from the sender:
@@ -396,46 +173,144 @@ impl Signaling {
         // Otherwise, the peer:
         //
         // * MUST ensure that the 16 byte cookie of the sender has not changed
-        {
-            let cookie_pair = peer.cookie_pair_mut();
-            match cookie_pair.theirs {
-                None => {
-                    // This is the first message from that peer,
+        let role = self.role();
+        let peer: &mut PeerContext = self.get_peer_with_address_mut(nonce.source()).ok_or_else(|| {
+            if role == Role::Initiator && nonce.source().is_responder() {
+                ValidationError::Fail(format!("Could not find responder with address {}", nonce.source()))
+            } else {
+                ValidationError::Crash("Got message from invalid sender that wasn't dropped".into())
+            }
+        })?;
+
+        let peer_identity = peer.identity();
+        let cookie_pair = peer.cookie_pair_mut();
+
+        match cookie_pair.theirs {
+            None => {
+                // This is the first message from that peer,
+                if *nonce.cookie() == cookie_pair.ours {
                     // validate the cookie...
-                    if *nonce.cookie() == cookie_pair.ours {
-                        let msg = format!("Cookie from {} is identical to our own cookie", peer_identity);
-                        return ValidationResult::Fail(msg);
-                    }
+                    Err(ValidationError::Fail(
+                        format!("Cookie from {} is identical to our own cookie", peer_identity)
+                    ))
+                } else  {
                     // ...and store it.
                     cookie_pair.theirs = Some(nonce.cookie().clone());
-                },
-                Some(ref cookie) => {
-                    // Ensure that the cookie has not changed
-                    if nonce.cookie() != cookie {
-                        let msg = format!("Cookie from {} has changed", peer_identity);
-                        return ValidationResult::Fail(msg);
-                    }
-                },
-            }
+                    Ok(())
+                }
+            },
+            Some(ref cookie) => {
+                // Ensure that the cookie has not changed
+                if nonce.cookie() != cookie {
+                    Err(ValidationError::Fail(
+                        format!("Cookie from {} has changed", peer_identity)
+                    ))
+                } else {
+                    Ok(())
+                }
+            },
         }
-
-        ValidationResult::Ok
     }
 
+    /// Handle an incoming message.
+    fn handle_message(&mut self, bbox: ByteBox) -> SignalingResult<Vec<HandleAction>> {
+        // Validate the nonce
+        match self.validate_nonce(&bbox.nonce) {
+            // It's valid! Carry on.
+            Ok(_) => {},
+
+            // Drop and ignore some of the messages
+            Err(ValidationError::DropMsg(warning)) => {
+                warn!("Invalid nonce: {}", warning);
+                return Ok(vec![]);
+            },
+
+            // Nonce is invalid, fail the signaling
+            Err(ValidationError::Fail(reason)) =>
+                return Err(SignalingError::InvalidNonce(reason)),
+
+            // A critical error occurred
+            Err(ValidationError::Crash(reason)) =>
+                return Err(SignalingError::Crash(reason)),
+        };
+
+        // Decode message depending on source
+        let obox: OpenBox = if bbox.nonce.source().is_server() {
+            self.decode_server_message(bbox)?
+        } else {
+            self.decode_peer_message(bbox)?
+        };
+
+        // Handle message depending on state
+        match self.common().signaling_state() {
+            SignalingState::ServerHandshake =>
+                self.handle_server_message(obox),
+
+            SignalingState::PeerHandshake if obox.nonce.source().is_server() =>
+                self.handle_server_message(obox),
+            SignalingState::PeerHandshake =>
+                self.handle_peer_message(obox),
+
+            SignalingState::Task =>
+                unimplemented!("TODO: Handle task messages"),
+        }
+    }
+
+
+    // Message decoding
+
     /// Decode or decrypt a binary message coming from the server
-    fn decode_server_msg(&self, bbox: ByteBox) -> SignalingResult<OpenBox> {
+    fn decode_server_message(&self, bbox: ByteBox) -> SignalingResult<OpenBox> {
         // The very first message from the server is unencrypted
-        if self.signaling_state() == SignalingState::ServerHandshake
-        && self.server().handshake_state() == ServerHandshakeState::New {
+        if self.common().signaling_state() == SignalingState::ServerHandshake
+        && self.server_handshake_state() == ServerHandshakeState::New {
             return bbox.decode();
         }
 
         // Otherwise, decrypt with server key
         match self.server().permanent_key {
-            Some(ref pubkey) => bbox.decrypt(&self.permanent_key(), pubkey),
+            Some(ref pubkey) => bbox.decrypt(&self.common().permanent_key, pubkey),
             None => Err(SignalingError::Crash("Missing server permanent key".into())),
         }
     }
+
+    /// Decrypt a binary message coming from a peer
+    fn decode_peer_message(&self, bbox: ByteBox) -> SignalingResult<OpenBox>;
+
+
+    // Message handling: Dispatching
+
+    /// Determine the next server handshake state based on the incoming
+    /// server-to-client message and the current state.
+    ///
+    /// This method call may have some side effects, like updates in the peer
+    /// context (cookie, CSN, etc).
+    fn handle_server_message(&mut self, obox: OpenBox) -> SignalingResult<Vec<HandleAction>> {
+        let old_state = self.server_handshake_state().clone();
+        match (old_state, obox.message) {
+            // Valid state transitions
+            (ServerHandshakeState::New, Message::ServerHello(msg)) =>
+                self.handle_server_hello(msg),
+            (ServerHandshakeState::ClientInfoSent, Message::ServerAuth(msg)) =>
+                self.handle_server_auth(msg),
+            (ServerHandshakeState::Done, Message::NewResponder(msg)) =>
+                self.handle_new_responder(msg),
+            (ServerHandshakeState::Done, Message::DropResponder(_msg)) =>
+                unimplemented!("Handling DropResponder messages not yet implemented"),
+            (ServerHandshakeState::Done, Message::SendError(msg)) =>
+                self.handle_send_error(msg),
+
+            // Any undefined state transition results in an error
+            (s, message) => Err(SignalingError::InvalidStateTransition(
+                format!("Got {} message from server in {:?} state", message.get_type(), s)
+            )),
+        }
+    }
+
+    fn handle_peer_message(&mut self, obox: OpenBox) -> SignalingResult<Vec<HandleAction>>;
+
+
+    // Message handling: Handling
 
     /// Handle an incoming [`ServerHello`](messages/struct.ServerHello.html) message.
     fn handle_server_hello(&mut self, msg: ServerHello) -> SignalingResult<Vec<HandleAction>> {
@@ -450,19 +325,19 @@ impl Signaling {
                 "Got a server-hello message, but server permanent key is already set".to_string()
             ));
         }
-        self.server_mut().permanent_key = Some(msg.key);
+        self.common_mut().server.permanent_key = Some(msg.key);
 
         // Reply with client-hello message if we're a responder
         if self.role() == Role::Responder {
             let client_hello = {
-                let key = self.permanent_key().public_key();
+                let key = self.common().permanent_key.public_key();
                 ClientHello::new(*key).into_message()
             };
             let client_hello_nonce = Nonce::new(
                 // Cookie
                 self.server().cookie_pair().ours.clone(),
                 // Src
-                self.identity().into(),
+                self.common().identity.into(),
                 // Dst
                 self.server().identity().into(),
                 // Csn
@@ -490,7 +365,7 @@ impl Signaling {
         match self.server().permanent_key {
             Some(ref pubkey) => {
                 debug!("<-- Enqueuing client-auth");
-                actions.push(HandleAction::Reply(reply.encrypt(&self.permanent_key(), pubkey)));
+                actions.push(HandleAction::Reply(reply.encrypt(&self.common().permanent_key, pubkey)));
             },
             None => return Err(SignalingError::Crash("Missing server permanent key".into())),
         };
@@ -536,96 +411,103 @@ impl Signaling {
         // TODO: Implement
 
         // Moreover, the client MUST do some checks depending on its role
-        let actions = on_inner!(self, ref mut s, s.handle_server_auth(&msg))?;
+        let actions = self.handle_server_auth_impl(&msg)?;
 
         info!("Server handshake completed");
         self.server_mut().set_handshake_state(ServerHandshakeState::Done);
-        self.set_signaling_state(SignalingState::PeerHandshake)?;
+        self.common_mut().set_signaling_state(SignalingState::PeerHandshake)?;
         Ok(actions)
     }
+
+    fn handle_server_auth_impl(&mut self, msg: &ServerAuth) -> SignalingResult<Vec<HandleAction>>;
+
+    /// Handle an incoming [`NewResponder`](messages/struct.NewResponder.html) message.
+    fn handle_new_responder(&mut self, msg: NewResponder) -> SignalingResult<Vec<HandleAction>>;
 
     /// Handle an incoming [`SendError`](messages/struct.ServerAuth.html) message.
     fn handle_send_error(&mut self, msg: SendError) -> SignalingResult<Vec<HandleAction>> {
         warn!("--> Received send-error");
-        trace!("Message that could not be relayed: {:?}", msg.id);
+        debug!("Message that could not be relayed: {:#?}", msg.id);
         return Err(SignalingError::SendError);
     }
-
-    /// Return the inner `InitiatorSignaling` instance.
-    ///
-    /// Panics if we're not an initiator
-    #[cfg(test)]
-    fn as_initiator(&self) -> &InitiatorSignaling {
-        match *self {
-            Signaling::Initiator(ref s) => s,
-            Signaling::Responder(_) => panic!("Called .as_initiator() on a `Signaling::Responder`"),
-        }
-    }
-
-    /// Return the inner `InitiatorSignaling` instance.
-    ///
-    /// Panics if we're not an initiator
-    #[cfg(test)]
-    fn as_initiator_mut(&mut self) -> &mut InitiatorSignaling {
-        match *self {
-            Signaling::Initiator(ref mut s) => s,
-            Signaling::Responder(_) => panic!("Called .as_initiator() on a `Signaling::Responder`"),
-        }
-    }
-
-    /// Return the inner `ResponderSignaling` instance.
-    ///
-    /// Panics if we're not an responder
-    #[cfg(test)]
-    fn as_responder(&self) -> &ResponderSignaling {
-        match *self {
-            Signaling::Responder(ref s) => s,
-            Signaling::Initiator(_) => panic!("Called .as_responder() on a `Signaling::Initiator`"),
-        }
-    }
-
-    /// Return the inner `ResponderSignaling` instance.
-    ///
-    /// Panics if we're not an responder
-    #[cfg(test)]
-    fn as_responder_mut(&mut self) -> &mut ResponderSignaling {
-        match *self {
-            Signaling::Responder(ref mut s) => s,
-            Signaling::Initiator(_) => panic!("Called .as_responder() on a `Signaling::Initiator`"),
-        }
-    }
 }
 
-
-/// Result of the nonce validation.
-pub(crate) enum ValidationResult {
-    /// Validation passed
-    Ok,
-    /// Ignore message
-    DropMsg(String),
-    /// Validation failed
-    Fail(String),
-    /// A critical error occurred
-    Crash(String),
-}
-
-
-/// Signaling data for the initiator.
-pub(crate) struct InitiatorSignaling {
+/// Common functionality / state for all signaling types.
+pub(crate) struct Common {
     // The signaling state
-    pub(crate) signaling_state: SignalingState,
+    signaling_state: SignalingState,
 
     // Our permanent keypair
     pub(crate) permanent_key: KeyStore,
 
+    // Our session keypair
+    pub(crate) session_key: Option<KeyStore>,
+
     // An optional auth token
     pub(crate) auth_token: Option<AuthToken>,
+
+    // The assigned role
+    pub(crate) role: Role,
 
     // The assigned client identity
     pub(crate) identity: ClientIdentity,
 
     // The server context
     pub(crate) server: ServerContext,
+}
+
+impl Common {
+    /// Return the current signaling state.
+    fn signaling_state(&self) -> SignalingState {
+        self.signaling_state
+    }
+
+    /// Set the current signaling state.
+    fn set_signaling_state(&mut self, state: SignalingState) -> SignalingResult<()> {
+        if self.signaling_state == state {
+            trace!("Ignoring signaling state transition: {:?} -> {:?}", self.signaling_state(), state);
+            return Ok(())
+        }
+        if !self.signaling_state.may_transition_to(state) {
+            return Err(SignalingError::InvalidStateTransition(
+                format!("Signaling state: {:?} -> {:?}", self.signaling_state(), state)
+            ));
+        }
+        trace!("Signaling state transition: {:?} -> {:?}", self.signaling_state(), state);
+        self.signaling_state = state;
+        Ok(())
+    }
+
+    /// Generate a session key.
+    fn generate_session_key(&mut self) -> SignalingResult<()> {
+        if self.session_key.is_some() {
+            return Err(
+                SignalingError::Crash("Cannot generate new session key: It has already been generated".into())
+            );
+        }
+
+        // The client MUST generate a session key pair (a new NaCl key pair for
+        // public key authenticated encryption) for further communication with
+        // the other client.
+        //
+        // Note: This *could* cause a panic if libsodium initialization fails, but
+        // that's not possible in practice because libsodium should already
+        // have been initialized previously.
+        let mut session_key = KeyStore::new().expect("Libsodium initialization failed");
+        while session_key == self.permanent_key {
+            warn!("Session keypair == permanent keypair! This is highly unlikely. Regenerating...");
+            session_key = KeyStore::new().expect("Libsodium initialization failed");
+        }
+        self.session_key = Some(session_key);
+        Ok(())
+    }
+}
+
+
+/// Signaling data for the initiator.
+pub(crate) struct InitiatorSignaling {
+    // Common state and functionality
+    pub(crate) common: Common,
 
     // The list of responders
     pub(crate) responders: HashMap<Address, ResponderContext>,
@@ -634,21 +516,83 @@ pub(crate) struct InitiatorSignaling {
     pub(crate) responder: Option<ResponderContext>,
 }
 
-impl InitiatorSignaling {
-    pub(crate) fn new(permanent_key: KeyStore) -> Self {
-        InitiatorSignaling {
-            signaling_state: SignalingState::ServerHandshake,
-            identity: ClientIdentity::Unknown,
-            server: ServerContext::new(),
-            permanent_key: permanent_key,
-            auth_token: Some(AuthToken::new()),
-            responders: HashMap::new(),
-            responder: None,
+impl Signaling for InitiatorSignaling {
+    /// Return a reference to the `Common` struct.
+    fn common(&self) -> &Common {
+        &self.common
+    }
+
+    /// Return a mutable reference to the `Common` struct.
+    fn common_mut(&mut self) -> &mut Common {
+        &mut self.common
+    }
+
+    fn get_peer(&self) -> Option<&PeerContext> {
+        self.responder.as_ref().map(|p| p as &PeerContext)
+    }
+
+    fn get_peer_with_address_mut(&mut self, addr: Address) -> Option<&mut PeerContext> {
+        let identity: Identity = addr.into();
+        match identity {
+            Identity::Server => Some(&mut self.common.server as &mut PeerContext),
+            Identity::Initiator => None,
+            Identity::Responder(_) => self.responders.get_mut(&addr).map(|r| r as &mut PeerContext),
+            Identity::Unknown => unreachable!(),
         }
     }
 
-    /// Decode or decrypt a binary message coming from the peer
-    fn decode_peer_msg(&self, bbox: ByteBox) -> SignalingResult<OpenBox> {
+    fn validate_nonce_destination(&mut self, nonce: &Nonce) -> Result<(), ValidationError> {
+		// A client MUST check that the destination address targets its
+		// assigned identity (or `0x00` during authentication).
+        if self.identity() == ClientIdentity::Unknown
+        && !nonce.destination().is_unknown()
+        && self.server_handshake_state() != ServerHandshakeState::New {
+            // The first message received with a destination address different
+            // to `0x00` SHALL be accepted as the client's assigned identity.
+            // However, the client MUST validate that the identity fits its
+            // role – initiators SHALL ONLY accept `0x01`. The identity MUST
+            // be stored as the client's assigned identity.
+            if nonce.destination().is_initiator() {
+                self.common.identity = ClientIdentity::Initiator;
+                debug!("Assigned identity: {}", self.identity());
+            } else {
+                return Err(ValidationError::Fail(
+                    format!("cannot assign address {} to initiator", nonce.destination())
+                ));
+            };
+        }
+        if nonce.destination() != self.identity().into() {
+            return Err(ValidationError::Fail(
+                format!("Bad destination: {} (our identity is {})", nonce.destination(), self.identity())
+            ));
+        }
+
+        Ok(())
+    }
+
+    fn validate_nonce_source(&mut self, nonce: &Nonce) -> Result<(), ValidationError> {
+        // An initiator SHALL ONLY process messages from the server (0x00). As
+        // soon as the initiator has been assigned an identity, it MAY ALSO accept
+        // messages from other responders (0x02..0xff). Other messages SHALL be
+        // discarded and SHOULD trigger a warning.
+        match nonce.source() {
+            // From server
+            Address(0x00) => Ok(()),
+
+            // From initiator
+            Address(0x01) => Err(ValidationError::DropMsg(
+                format!("Bad source: {} (our identity is {})", nonce.source(), self.identity())
+            )),
+
+            // From responder
+            Address(0x02...0xff) => Ok(()),
+
+            // Required due to https://github.com/rust-lang/rfcs/issues/1550
+            Address(_) => unreachable!(),
+        }
+    }
+
+    fn decode_peer_message(&self, bbox: ByteBox) -> SignalingResult<OpenBox> {
         // Validate source again
         if !bbox.nonce.source().is_responder() {
             return Err(SignalingError::Crash(format!("Received message from an initiator")));
@@ -677,7 +621,7 @@ impl InitiatorSignaling {
         match responder.handshake_state() {
             ResponderHandshakeState::New => {
                 // Expect token message, encrypted with authentication token.
-                match self.auth_token {
+                match self.common.auth_token {
                     Some(ref token) => bbox.decrypt_token(token),
                     None => Err(SignalingError::Crash("Auth token not set".into())),
                 }
@@ -685,7 +629,7 @@ impl InitiatorSignaling {
             ResponderHandshakeState::TokenReceived => {
                 // Expect key message, encrypted with our public permanent key
                 // and responder private permanent key
-                bbox.decrypt(&self.permanent_key, responder_permanent_key(&responder)?)
+                bbox.decrypt(&self.common.permanent_key, responder_permanent_key(&responder)?)
             },
             ResponderHandshakeState::KeySent => {
                 // Expect auth message, encrypted with our public session key
@@ -728,7 +672,7 @@ impl InitiatorSignaling {
         }
     }
 
-    fn handle_server_auth(&mut self, msg: &ServerAuth) -> SignalingResult<Vec<HandleAction>> {
+    fn handle_server_auth_impl(&mut self, msg: &ServerAuth) -> SignalingResult<Vec<HandleAction>> {
         // In case the client is the initiator, it SHALL check
         // that the responders field is set and contains an
         // Array of responder identities.
@@ -781,7 +725,6 @@ impl InitiatorSignaling {
         Ok(vec![])
     }
 
-    /// Handle an incoming [`NewResponder`](messages/struct.NewResponder.html) message.
     fn handle_new_responder(&mut self, msg: NewResponder) -> SignalingResult<Vec<HandleAction>> {
         debug!("--> Received new-responder");
 
@@ -809,6 +752,25 @@ impl InitiatorSignaling {
         // TODO: Implement
 
         Ok(vec![])
+    }
+
+}
+
+impl InitiatorSignaling {
+    pub(crate) fn new(permanent_key: KeyStore) -> Self {
+        InitiatorSignaling {
+            common: Common {
+                signaling_state: SignalingState::ServerHandshake,
+                role: Role::Initiator,
+                identity: ClientIdentity::Unknown,
+                permanent_key: permanent_key,
+                session_key: None,
+                auth_token: Some(AuthToken::new()),
+                server: ServerContext::new(),
+            },
+            responders: HashMap::new(),
+            responder: None,
+        }
     }
 
     /// Handle an incoming [`Token`](messages/struct.Token.html) message.
@@ -873,13 +835,13 @@ impl InitiatorSignaling {
         }.into_message();
         let key_nonce = Nonce::new(
             responder.cookie_pair().ours.clone(),
-            self.identity.into(),
+            self.common.identity.into(),
             responder.identity().into(),
             responder.csn_pair().borrow_mut().ours.increment()?,
         );
         let obox = OpenBox::new(key, key_nonce);
         let bbox = obox.encrypt(
-            &self.permanent_key,
+            &self.common.permanent_key,
             responder.permanent_key.as_ref()
                 .ok_or(SignalingError::Crash("Responder permanent key not set".into()))?,
         );
@@ -887,50 +849,96 @@ impl InitiatorSignaling {
         debug!("<-- Enqueuing key");
         Ok(vec![HandleAction::Reply(bbox)])
     }
-
 }
+
 
 /// Signaling data for the responder.
 pub(crate) struct ResponderSignaling {
-    // The signaling state
-    pub(crate) signaling_state: SignalingState,
-
-    // Our permanent keypair
-    pub(crate) permanent_key: KeyStore,
-
-    // Our session keypair
-    pub(crate) session_key: Option<KeyStore>,
-
-    // An optional auth token
-    pub(crate) auth_token: Option<AuthToken>,
-
-    // The assigned client identity
-    pub(crate) identity: ClientIdentity,
-
-    // The server context
-    pub(crate) server: ServerContext,
+    // Common state and functionality
+    pub(crate) common: Common,
 
     // The initiator context
     pub(crate) initiator: InitiatorContext,
 }
 
-impl ResponderSignaling {
-    pub(crate) fn new(permanent_key: KeyStore,
-                      initiator_pubkey: PublicKey,
-                      auth_token: Option<AuthToken>) -> Self {
-        ResponderSignaling {
-            signaling_state: SignalingState::ServerHandshake,
-            permanent_key: permanent_key,
-            session_key: None,
-            auth_token: auth_token,
-            identity: ClientIdentity::Unknown,
-            server: ServerContext::new(),
-            initiator: InitiatorContext::new(initiator_pubkey),
+impl Signaling for ResponderSignaling {
+    /// Return a reference to the `Common` struct.
+    fn common(&self) -> &Common {
+        &self.common
+    }
+
+    /// Return a mutable reference to the `Common` struct.
+    fn common_mut(&mut self) -> &mut Common {
+        &mut self.common
+    }
+
+    fn get_peer(&self) -> Option<&PeerContext> {
+        Some(&self.initiator as &PeerContext)
+    }
+
+    fn get_peer_with_address_mut(&mut self, addr: Address) -> Option<&mut PeerContext> {
+        let identity: Identity = addr.into();
+        match identity {
+            Identity::Server => Some(&mut self.common.server),
+            Identity::Initiator => Some(&mut self.initiator),
+            Identity::Responder(_) => None,
+            Identity::Unknown => unreachable!(),
         }
     }
 
-    /// Decode or decrypt a binary message coming from the peer
-    fn decode_peer_msg(&self, bbox: ByteBox) -> SignalingResult<OpenBox> {
+    fn validate_nonce_destination(&mut self, nonce: &Nonce) -> Result<(), ValidationError> {
+		// A client MUST check that the destination address targets its
+		// assigned identity (or `0x00` during authentication).
+        if self.identity() == ClientIdentity::Unknown
+        && !nonce.destination().is_unknown()
+        && self.server_handshake_state() != ServerHandshakeState::New {
+            // The first message received with a destination address different
+            // to `0x00` SHALL be accepted as the client's assigned identity.
+            // However, the client MUST validate that the identity fits its
+            // role – responders SHALL ONLY an identity from the range
+            // `0x02..0xff`. The identity MUST be stored as the client's
+            // assigned identity.
+            if nonce.destination().is_responder() {
+                self.common.identity = ClientIdentity::Responder(nonce.destination().0);
+                debug!("Assigned identity: {}", self.identity());
+            } else {
+                return Err(ValidationError::Fail(
+                    format!("cannot assign address {} to a responder", nonce.destination())
+                ));
+            };
+        }
+        if nonce.destination() != self.identity().into() {
+            return Err(ValidationError::Fail(
+                format!("Bad destination: {} (our identity is {})", nonce.destination(), self.identity())
+            ));
+        }
+
+        Ok(())
+    }
+
+    fn validate_nonce_source(&mut self, nonce: &Nonce) -> Result<(), ValidationError> {
+        // A responder SHALL ONLY process messages from the server (0x00). As soon
+        // as the responder has been assigned an identity, it MAY ALSO accept
+        // messages from the initiator (0x01). Other messages SHALL be discarded
+        // and SHOULD trigger a warning.
+        match nonce.source() {
+            // From server
+            Address(0x00) => Ok(()),
+
+            // From initiator
+            Address(0x01) => Ok(()),
+
+            // From responder
+            Address(0x02...0xff) => Err(ValidationError::DropMsg(
+                format!("Bad source: {} (our identity is {})", nonce.source(), self.identity())
+            )),
+
+            // Required due to https://github.com/rust-lang/rfcs/issues/1550
+            Address(_) => unreachable!(),
+        }
+    }
+
+    fn decode_peer_message(&self, bbox: ByteBox) -> SignalingResult<OpenBox> {
         // Validate source again
         if !bbox.nonce.source().is_initiator() {
             return Err(SignalingError::Crash(format!("Received message from a responder")));
@@ -941,12 +949,12 @@ impl ResponderSignaling {
             InitiatorHandshakeState::KeySent => {
                 // Expect key message, encrypted with our public permanent key
                 // and initiator private permanent key
-                bbox.decrypt(&self.permanent_key, &self.initiator.permanent_key)
+                bbox.decrypt(&self.common.permanent_key, &self.initiator.permanent_key)
             },
             InitiatorHandshakeState::AuthSent => {
                 // Expect an auth message, encrypted with our public session
                 // key and initiator private session key
-                match (self.session_key.as_ref(), self.initiator.session_key.as_ref()) {
+                match (self.common().session_key.as_ref(), self.initiator.session_key.as_ref()) {
                     (Some(ref our_key), Some(ref initiator_key)) => bbox.decrypt(our_key, initiator_key),
                     (None, _) => return Err(SignalingError::Crash("Our session key not set".into())),
                     (_, None) => return Err(SignalingError::Crash("Initiator session key not set".into())),
@@ -977,7 +985,7 @@ impl ResponderSignaling {
         }
     }
 
-    fn handle_server_auth(&mut self, msg: &ServerAuth) -> SignalingResult<Vec<HandleAction>> {
+    fn handle_server_auth_impl(&mut self, msg: &ServerAuth) -> SignalingResult<Vec<HandleAction>> {
         // In case the client is the responder, it SHALL check
         // that the initiator_connected field contains a
         // boolean value.
@@ -989,12 +997,12 @@ impl ResponderSignaling {
         let mut actions: Vec<HandleAction> = vec![];
         match msg.initiator_connected {
             Some(true) => {
-                if let Some(ref token) = self.auth_token {
+                if let Some(ref token) = self.common().auth_token {
                     actions.push(self.send_token(&token)?);
                 } else {
                     debug!("No auth token set");
                 }
-                self.generate_session_key()?;
+                self.common_mut().generate_session_key()?;
                 actions.push(self.send_key()?);
                 self.initiator.set_handshake_state(InitiatorHandshakeState::KeySent);
             },
@@ -1008,32 +1016,27 @@ impl ResponderSignaling {
         Ok(actions)
     }
 
-    /// Handle an incoming [`NewResponder`](messages/struct.NewResponder.html) message.
     fn handle_new_responder(&mut self, _msg: NewResponder) -> SignalingResult<Vec<HandleAction>> {
         Err(SignalingError::Protocol("Received 'new-responder' message as responder".into()))
     }
+}
 
-    fn generate_session_key(&mut self) -> SignalingResult<()> {
-        if self.session_key.is_some() {
-            return Err(
-                SignalingError::Crash("Cannot generate new session key: It has already been generated".into())
-            );
+impl ResponderSignaling {
+    pub(crate) fn new(permanent_key: KeyStore,
+                      initiator_pubkey: PublicKey,
+                      auth_token: Option<AuthToken>) -> Self {
+        ResponderSignaling {
+            common: Common {
+                signaling_state: SignalingState::ServerHandshake,
+                role: Role::Responder,
+                identity: ClientIdentity::Unknown,
+                permanent_key: permanent_key,
+                session_key: None,
+                auth_token: auth_token,
+                server: ServerContext::new(),
+            },
+            initiator: InitiatorContext::new(initiator_pubkey),
         }
-
-        // The client MUST generate a session key pair (a new NaCl key pair for
-        // public key authenticated encryption) for further communication with
-        // the other client.
-        //
-        // Note: This *could* cause a panic if libsodium initialization fails, but
-        // that's not possible in practice because libsodium should already
-        // have been initialized previously.
-        let mut session_key = KeyStore::new().expect("Libsodium initialization failed");
-        while session_key == self.permanent_key {
-            warn!("Session keypair == permanent keypair! This is highly unlikely. Regenerating...");
-            session_key = KeyStore::new().expect("Libsodium initialization failed");
-        }
-        self.session_key = Some(session_key);
-        Ok(())
     }
 
     /// Build a `Token` message.
@@ -1041,11 +1044,11 @@ impl ResponderSignaling {
         // The responder MUST set the public key (32 bytes) of the permanent
         // key pair in the key field of this message.
         let msg: Message = Token {
-            key: self.permanent_key.public_key().to_owned(),
+            key: self.common().permanent_key.public_key().to_owned(),
         }.into_message();
         let nonce = Nonce::new(
             self.initiator.cookie_pair().ours.clone(),
-            self.identity.into(),
+            self.identity().into(),
             self.initiator.identity().into(),
             self.initiator.csn_pair().borrow_mut().ours.increment()?,
         );
@@ -1066,7 +1069,7 @@ impl ResponderSignaling {
     /// Build a `Key` message.
     fn send_key(&self) -> SignalingResult<HandleAction> {
         // It MUST set the public key (32 bytes) of that key pair in the key field.
-        let msg: Message = match self.session_key {
+        let msg: Message = match self.common().session_key {
             Some(ref session_key) => Key {
                 key: session_key.public_key().to_owned(),
             }.into_message(),
@@ -1074,7 +1077,7 @@ impl ResponderSignaling {
         };
         let nonce = Nonce::new(
             self.initiator.cookie_pair().ours.clone(),
-            self.identity.into(),
+            self.identity().into(),
             self.initiator.identity().into(),
             self.initiator.csn_pair().borrow_mut().ours.increment()?,
         );
@@ -1082,7 +1085,7 @@ impl ResponderSignaling {
 
         // The message SHALL be NaCl public-key encrypted by the client's
         // permanent key pair and the other client's permanent key pair.
-        let bbox = obox.encrypt(&self.permanent_key, &self.initiator.permanent_key);
+        let bbox = obox.encrypt(&self.common().permanent_key, &self.initiator.permanent_key);
 
         debug!("<-- Enqueuing key");
         Ok(HandleAction::Reply(bbox))
@@ -1115,13 +1118,13 @@ impl ResponderSignaling {
             .into_message();
         let auth_nonce = Nonce::new(
             self.initiator.cookie_pair().ours.clone(),
-            self.identity.into(),
+            self.common().identity.into(),
             self.initiator.identity().into(),
             self.initiator.csn_pair().borrow_mut().ours.increment()?,
         );
         let obox = OpenBox::new(auth, auth_nonce);
         let bbox = obox.encrypt(
-            self.session_key.as_ref()
+            self.common().session_key.as_ref()
                 .ok_or(SignalingError::Crash("Our session key not set".into()))?,
             self.initiator.session_key.as_ref()
                 .ok_or(SignalingError::Crash("Initiator session key not set".into()))?,
@@ -1130,6 +1133,17 @@ impl ResponderSignaling {
         debug!("<-- Enqueuing auth");
         Ok(vec![HandleAction::Reply(bbox)])
     }
+}
+
+
+/// Result of the nonce validation.
+pub(crate) enum ValidationError {
+    /// Ignore message
+    DropMsg(String),
+    /// Validation failed
+    Fail(String),
+    /// A critical error occurred
+    Crash(String),
 }
